Junit5 added, Project lombock added, small fixes, rename groupId
