--- backup_pom_test_config
@@ -1,101 +1,0 @@
-<dependency>
-			<groupId>org.testcontainers</groupId>
-			<artifactId>junit-jupiter</artifactId>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.testcontainers</groupId>
-			<artifactId>mysql</artifactId>
-			<scope>test</scope>
-		</dependency>
-
-
-
-
-
-<dependencyManagement>
-		<dependencies>
-			<dependency>
-				<groupId>org.testcontainers</groupId>
-				<artifactId>testcontainers-bom</artifactId>
-				<version>${testcontainers.version}</version>
-				<type>pom</type>
-				<scope>import</scope>
-			</dependency>
-		</dependencies>
-	</dependencyManagement>
-
-
-
-	BackUP POM
-
-
-
-	<?xml version="1.0" encoding="UTF-8"?>
-    <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    	<modelVersion>4.0.0</modelVersion>
-
-    	<groupId>com.staxrt.tutorial</groupId>
-    	<artifactId>spring-boot-rest-api-tutorial</artifactId>
-    	<version>0.0.1-SNAPSHOT</version>
-    	<packaging>jar</packaging>
-
-    	<name>spring-boot-rest-api-tutorial</name>
-    	<description>Sample REST CRUD API project using Spring Boot, Mysql, JPA and Hibernate</description>
-
-    	<parent>
-    		<groupId>org.springframework.boot</groupId>
-    		<artifactId>spring-boot-starter-parent</artifactId>
-    		<version>2.1.2.RELEASE</version>
-    		<relativePath/> <!-- lookup parent from repository -->
-    	</parent>
-
-    	<properties>
-    		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-    		<java.version>1.8</java.version>
-    	</properties>
-
-    	<dependencies>
-    		<dependency>
-    			<groupId>org.springframework.boot</groupId>
-    			<artifactId>spring-boot-starter-actuator</artifactId>
-    		</dependency>
-    		<dependency>
-    			<groupId>org.springframework.boot</groupId>
-    			<artifactId>spring-boot-starter-data-jpa</artifactId>
-    		</dependency>
-    		<dependency>
-    			<groupId>org.springframework.boot</groupId>
-    			<artifactId>spring-boot-starter-web</artifactId>
-    		</dependency>
-
-    		<dependency>
-    			<groupId>org.springframework.boot</groupId>
-    			<artifactId>spring-boot-devtools</artifactId>
-    			<scope>runtime</scope>
-    		</dependency>
-    		<dependency>
-    			<groupId>mysql</groupId>
-    			<artifactId>mysql-connector-java</artifactId>
-    			<scope>runtime</scope>
-    		</dependency>
-    		<dependency>
-    			<groupId>org.springframework.boot</groupId>
-    			<artifactId>spring-boot-starter-test</artifactId>
-    			<scope>test</scope>
-    		</dependency>
-    	</dependencies>
-
-    	<build>
-    		<plugins>
-    			<plugin>
-    				<groupId>org.springframework.boot</groupId>
-    				<artifactId>spring-boot-maven-plugin</artifactId>
-    			</plugin>
-    		</plugins>
-    	</build>
-
-
-    </project>
