+++ src/main/java/com/github/elmo675/exception/ErrorResponse.java
@@ -1,6 +1,6 @@
 
 
-package com.api.v1.exception;
+package com.github.elmo675.exception;
 
 import java.util.Date;
