+++ src/main/java/com/github/elmo675/exception/ResourceNotFoundException.java
@@ -1,5 +1,5 @@
 
-package com.api.v1.exception;
+package com.github.elmo675.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
