+++ src/main/java/com/github/elmo675/model/DiaryEntry.java
@@ -1,0 +1,50 @@
+package com.github.elmo675.model;
+
+import lombok.Data;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+import org.springframework.data.annotation.CreatedBy;
+import org.springframework.data.annotation.LastModifiedBy;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import javax.persistence.*;
+import java.util.Date;
+
+@Entity
+@Table(name = "DiaryEntries")
+@EntityListeners(AuditingEntityListener.class)
+@Data
+public class DiaryEntry {
+    enum publicity{Public,Friends,Private}
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private long id;
+
+
+    @Column(name = "content", nullable = false)
+    private String content;
+
+    @Column(name = "privacy", nullable = false)
+    private publicity privacy;
+
+
+    @CreationTimestamp
+    @Temporal(TemporalType.TIMESTAMP)
+    @Column(name = "created_at", nullable = false)
+    private Date createdAt;
+
+    @CreatedBy
+    @Column(name = "created_by", nullable = false)
+    private String createdBy;
+
+    @UpdateTimestamp
+    @Temporal(TemporalType.TIMESTAMP)
+    @Column(name = "updated_at", nullable = false)
+    private Date updatedAt;
+
+    @LastModifiedBy
+    @Column(name = "updated_by", nullable = false)
+    private String updatedBy;
+
+
+}
